Finished sign up page functionality
Added birthdate to the sign up page (including custom validation to check that the user is at least 16 years old as per the requirements), added user role to the sign up page (Instructor/Student), added password validation to the sign up page. All of these changes are reflected in the Account database model. Included my database Migrations in this commit.