Added Calendar button to navbar, added calendars to the calendar pages
Added Fullcalendar version 3.9.0 package to the project. Used the full calendar tutorial on the website to add the calendars to the calendar pages (separate pages for CalendarInstructor and CalendarStudent). Added the pageRole field to every page that can distinguish what role is assigned to that page to make it easier to navigate in the future. The calendar pages use this functionality with this commit. Added the field to the Index, SignUp2, WelcomeInstructor/Student, Profile, CalendarStudent/Instructor pages. If pageRole is null there is no role assigned to that page, if pageRole is false it is a student page, if pageRole is true it is an Instructor page.