Added checkboxes for MeetingDays selection in the course creation page
Added checkboxes for each day of the week when creating a course that will allow an instructor to select on which days of the week class will be held. Each checkbox is bound to a variable on the page model that is used to create the string of meeting days that is sent to the database for storage. If no days are selected, MeetingDays will show "None" in the database. Because of this, it is not required to select any meeting days when creating a course (for creating an asynchronous online course).