Added assignment due date functionality to the student calendar. Optimized the creation of class day events in both instructor and student calendars.
Changed the creation of class day events for students and instructors to create events that repeat so many days a week instead of several events for each day of the week that class occurs on. Requires one less loop of running code. Added assignment due dates to the calendar. Due to strange ISO8601 time shenanigans each event in the calendar lasts for 1 second to avoid 11:30pm due dates being stretched out on the calendar into the next day. The calendar is now feature complete (for now). More testing required for other assignment edge cases.