Updated Pages/Assignments/Index.cshtml to fix a bug where the create assignment button would not display when it should.
Changed the Pages/Assignments/Index.cshtml file to (when logged in as a professor for a course) first check if any assignments are due for a course. This will display "No assignments due." if there are no assignments for a course. Then if the user is the instructor for the course they will have the create new assignment button. Previously, the page would check for assignments and if there were none it would only display "No assignments due." and leave out the button and table entirely. Now each of them have separate checks one after another in the order -> no assignments due -> create assignment button -> assignment table (if any assignments exist).