Finished implementing a bar graph on student assignment page for each assignment submitted
Added an extra portion to the connection string that allows MARS (MultipleActiveResultSets=True;) to allow me to query the database an additional time (I would like to change this in the future but that will require learning how to close the database data readers which takes time I do not have right now). Also fixed a strange bug that would occur when logging in as a student. An invalid datetime object would appear with a date of 2100/5/5 or something like that.